+++ lib/taurus/core/evaluation/test/res/ipap_example.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python
-
 #############################################################################
 ##
 # This file is part of Taurus
@@ -23,8 +22,35 @@
 ##
 #############################################################################
 
-"""The sardana package. It contains specific part of sardana"""
+"""
+Examples on using the evaluation scheme for exposing icepap driver values
+as taurus attributes
+"""
 
-__docformat__ = 'restructuredtext'
 
-from formatter import tangoFormatter
+ATTR_IPAP_POS = ( 'eval:@ipap=pyIcePAP.EthIcePAP("icepap06", port=5000)' +
+                  '/float(ipap.readParameter(1,"POS"))')
+
+
+def _test1():
+    import taurus.core
+    a = taurus.Attribute(ATTR_IPAP_POS)
+    print "axis pos:", a.read().rvalue
+
+
+def _test2():
+    import sys
+    from taurus.qt.qtgui.application import TaurusApplication
+    from taurus.qt.qtgui.display import TaurusLabel
+    app = TaurusApplication()
+
+    tl = TaurusLabel()
+    tl.setModel(ATTR_IPAP_POS)
+    tl.show()
+
+    sys.exit(app.exec_())
+
+
+if __name__ == "__main__":
+    _test1()
+    _test2()