+++ lib/taurus/core/evaluation/test/test_evalvalidator.py
@@ -74,6 +74,8 @@ class EvaluationAuthValidatorTestCase(AbstractNameValidatorTestCase,
 @valid(name='eval:@foo')
 @valid(name='eval:@mymod.Myclass', groups={'devname': '@mymod.Myclass'})
 @valid(name='eval:@mymod.mysubmod.Myclass')
+@valid(name='eval:@mymod.mysubmod.*')
+@valid(name='eval:@mymod.*')
 @valid(name='eval:/@foo', groups={'devname': '@foo', 'path': '/@foo'})
 @valid(name='eval://localhost/@foo')
 @valid(name='eval:@ foo', groups={'devname': '@ foo'})
@@ -86,6 +88,16 @@ class EvaluationAuthValidatorTestCase(AbstractNameValidatorTestCase,
 @valid(name='eval:@foo')
 @invalid(name='eval://a/b', strict=True)
 @invalid(name='eval://d')
+# now ensure that attr names do not get validated as devs
+@invalid(name='eval:@mymod.MyClass()/c.foo')
+@invalid(name='eval:@c=mymod.MyClass(1)/c.foo()')
+@invalid(name='eval:@c=mymod.MyClass(1)/float(c.foo())')
+@invalid(name='eval:@c=mymod.MyClass("a/b",q=())/float(c.foo())')
+@valid(name='eval:@c=mymod.MyClass("a/b",q=())',
+       groups={'devname': '@c=mymod.MyClass("a/b",q=())'})
+@valid(name="eval:@c=mymod.MyClass('a/b',q=())",
+       groups={'devname': "@c=mymod.MyClass('a/b',q=())"})
+# check names
 @names(name='eval://localhost/@foo',
        out=('eval://localhost/@foo', '@foo', 'foo'))
 @names(name='eval://dev=foo',
@@ -112,6 +124,8 @@ class EvaluationDevValidatorTestCase(AbstractNameValidatorTestCase,
 @valid(name='eval:@foo/1')
 @valid(name='eval:@mymod.Myclass/1.2',
        groups={'attrname': '1.2', 'devname': '@mymod.Myclass'})
+@valid(name='eval:@mymod.*/bar',
+       groups={'attrname': 'bar', 'devname': '@mymod.*'})
 @valid(name='eval://linspace(-1, 1, 256)',
        groups={'attrname': 'linspace(-1, 1, 256)',
                '_expr': 'linspace(-1, 1, 256)',
@@ -310,6 +324,26 @@ class EvaluationDevValidatorTestCase(AbstractNameValidatorTestCase,
 @invalid(name='eval:{tango://a/b/c/d}')  # invalid because of non-strict ref
 # but valid with old syntax
 @valid(name='eval:{tango://a/b/c/d}', strict=False)
+#=========================================================================
+# Tests for eval attribute name validation (with custom evaluation dev)
+#=========================================================================
+@valid(name='eval:@mymod.*/foo')
+@valid(name='eval:@mymod.*/foo()')
+@valid(name='eval:@mymod.*/MyClass().foo()')
+@valid(name='eval:@mymod.*/foo()+1')
+@valid(name='eval:@mymod.*/foo()+{eval:@mymod.*/bar()}')
+@valid(name='eval:@c=mymod.MyClass()/c.foo',
+       groups={'devname': '@c=mymod.MyClass()',
+               'attrname':'c.foo'})
+@valid(name='eval:@c=mymod.MyClass()/c.foo()',
+       groups={'devname': '@c=mymod.MyClass()',
+               'attrname':'c.foo()'})
+@valid(name='eval:@c=mymod.MyClass(1)/float(c.foo())',
+       groups={'devname': '@c=mymod.MyClass(1)',
+               'attrname':'float(c.foo())'})
+@valid(name='eval:@c=mymod.MyClass("a/b",q=())/float(c.foo())',
+       groups={'devname': '@c=mymod.MyClass("a/b",q=())',
+               'attrname':'float(c.foo())'})
 #=========================================================================
 # Tests for eval attribute  name validation (when passing fragment / cfgkey)
 #=========================================================================