+++ lib/taurus/qt/qtgui/plot/taurustrend.py
@@ -298,8 +298,12 @@ class TaurusTrendsSet(Qt.QObject, TaurusBaseComponent):
         SPECTRUM attribute with dim_x=8. Then the return value will be (X,Y)
         where X.shape=(10,) and Y.shape=(10,8); X.dtype = Y.dtype = <dtype('float64')>
         '''
+        attr = self.getModelObj()
         if value is not None:
-            v = value.rvalue.magnitude  # TODO: check unit consistency
+            if attr.isNumeric():
+                v = value.rvalue.magnitude  # TODO: check unit consistency
+            else:
+                v = value.rvalue
             if numpy.isscalar(v):
                 ntrends = 1
             else:
@@ -324,11 +328,15 @@ class TaurusTrendsSet(Qt.QObject, TaurusBaseComponent):
             self._yBuffer = ArrayBuffer(numpy.zeros(
                 (min(128, self._maxBufferSize), ntrends), dtype='d'), maxSize=self._maxBufferSize)
         if value is not None:
+            if attr.isNumeric():
+                v = value.rvalue.magnitude
+            else:
+                v = value.rvalue
             try:
-                self._yBuffer.append(value.rvalue.magnitude)
+                self._yBuffer.append(v)
             except Exception, e:
                 self.warning('Problem updating history (%s=%s):%s',
-                             model, value.rvalue.magnitude, e)
+                             model, v, e)
                 value = None
 
         if self.parent().getXIsTime():
@@ -410,11 +418,14 @@ class TaurusTrendsSet(Qt.QObject, TaurusBaseComponent):
                     self._onDroppedEvent(reason='invalid value')
                     if not self.parent().getUseArchiving():
                         return
-                elif not hasattr(value.rvalue, 'magnitude'):
-                    self._onDroppedEvent(reason='rvalue has no .magnitude')
-                    return
+                elif model.isNumeric():
+                    if not hasattr(value.rvalue, 'magnitude'):
+                        self._onDroppedEvent(reason='rvalue has no .magnitude')
+                        return
+                    else:
+                        self._checkDataDimensions(value.rvalue.magnitude)
                 else:
-                    self._checkDataDimensions(value.rvalue.magnitude)
+                    self._checkDataDimensions(value.rvalue)
 
         # get the data from the event
         try:
@@ -467,7 +478,7 @@ class TaurusTrendsSet(Qt.QObject, TaurusBaseComponent):
 
         :param reason: (str) The reason of the drop
         '''
-        self.debug("Droping event. Reason %s", reason)
+        self.debug("Dropping event. Reason %s", reason)
         self.droppedEventsCount += 1
         self.consecutiveDroppedEventsCount += 1
         mustwarn = False